Address remaining review comments
