+++ packages/wonder-blocks-modal/src/components/__tests__/modal-backdrop.test.js
@@ -6,6 +6,7 @@ import ModalBackdrop from "../modal-backdrop.js";
 import OnePaneDialog from "../one-pane-dialog.js";
 
 import {unmountAll} from "../../../../../utils/testing/enzyme-shim.js";
+import {getElementAttachedToDocument} from "../../../../../utils/testing/get-element-attached-to-document.js";
 
 const wait = (duration: number = 0) =>
     new Promise((resolve, reject) => setTimeout(resolve, duration));
@@ -33,18 +34,6 @@ const exampleModalWithButtons = (
 );
 
 describe("ModalBackdrop", () => {
-    const getElementAttachedToDocument = (id: string): HTMLElement => {
-        const element = document.getElementById(id);
-        if (element) {
-            return element;
-        }
-
-        const newElement = document.createElement("div");
-        newElement.setAttribute("id", id);
-        document.body?.appendChild(newElement);
-        return newElement;
-    };
-
     beforeEach(() => {
         jest.useRealTimers();
     });
@@ -105,6 +94,8 @@ describe("ModalBackdrop", () => {
 
     test("If initialFocusId is set and element is found, we focus that element inside the modal", async () => {
         // Arrange
+        // We need the elements in the DOM document, it seems, for this test
+        // to work. Changing to testing-library will likely fix this.
         const attachElement = getElementAttachedToDocument("container");
         const initialFocusId = "initial-focus";
 
@@ -140,6 +131,8 @@ describe("ModalBackdrop", () => {
 
     test("If initialFocusId is set but element is NOT found, we focus on the first focusable element instead", async () => {
         // Arrange
+        // We need the elements in the DOM document, it seems, for this test
+        // to work. Changing to testing-library will likely fix this.
         const attachElement = getElementAttachedToDocument("container");
         const initialFocusId = "initial-focus";
         const firstFocusableElement = "[data-first-button]";
@@ -169,6 +162,8 @@ describe("ModalBackdrop", () => {
 
     test("If no initialFocusId is set, we focus the first button in the modal", async () => {
         // Arrange
+        // We need the elements in the DOM document, it seems, for this test
+        // to work. Changing to testing-library will likely fix this.
         const attachElement = getElementAttachedToDocument("container");
         const wrapper = mount(
             <ModalBackdrop onCloseModal={() => {}}>
@@ -189,6 +184,8 @@ describe("ModalBackdrop", () => {
 
     test("If there are no focusable elements, we focus the Dialog instead", async () => {
         // Arrange
+        // We need the elements in the DOM document, it seems, for this test
+        // to work. Changing to testing-library will likely fix this.
         const attachElement = getElementAttachedToDocument("container");
         const wrapper = mount(
             <ModalBackdrop onCloseModal={() => {}}>
